Initial support for aborting existing import.
Neo4j supports cancelling an existing import. Add in the action for
aborting and write in an optional flag to the start() method to force
aborting if we detect one is running.

This is mildly problematic as the Neo4j errors are vague. We can read
the message to determine if it was an existing import or there's already
a graph/database with the provided name...but for now let's keep it easy
in case that situation improves.